+++ src/main/scala/com/sfxcode/nosql/mongo/gridfs/Base.scala
@@ -1,12 +1,14 @@
 package com.sfxcode.nosql.mongo.gridfs
 
-import java.nio.channels.AsynchronousFileChannel
-import java.nio.file.{OpenOption, Path, StandardOpenOption}
+import java.io.OutputStream
+import java.nio.ByteBuffer
 
+import com.mongodb.client.gridfs.model.GridFSUploadOptions
+import com.sfxcode.nosql.mongo.Converter
 import com.typesafe.scalalogging.LazyLogging
 import org.mongodb.scala.bson.ObjectId
 import org.mongodb.scala.gridfs.{GridFSBucket, GridFSDownloadObservable}
-import org.mongodb.scala.{Observable, ReadConcern, ReadPreference, WriteConcern}
+import org.mongodb.scala.{Document, Observable, ReadConcern, ReadPreference, WriteConcern}
 
 abstract class Base extends LazyLogging {
 
@@ -20,10 +22,6 @@ abstract class Base extends LazyLogging {
     metadataKey
   }
 
-  def download(oid: ObjectId):GridFSDownloadObservable = {
-    gridfsBucket.downloadToObservable(oid)
-  }
-
   def drop(): Observable[Void] = gridfsBucket.drop()
 
   def bucketName: String = gridfsBucket.bucketName
@@ -35,5 +33,33 @@ abstract class Base extends LazyLogging {
   def readPreference: ReadPreference = gridfsBucket.readPreference
 
   def readConcern: ReadConcern = gridfsBucket.readConcern
+
+
+  def upload(fileName: String,
+             source: Observable[ByteBuffer],
+             metadata: AnyRef = Document(),
+             chunkSizeBytes: Int = 1024 * 256): Observable[ObjectId] = {
+    val metadataDocument = {
+      metadata match {
+        case document: Document => document
+        case _                  => Converter.toDocument(metadata)
+      }
+    }
+    val options: GridFSUploadOptions = new GridFSUploadOptions()
+      .chunkSizeBytes(chunkSizeBytes)
+      .metadata(metadataDocument)
+    gridfsBucket.uploadFromObservable(fileName, source, options)
+  }
+
+
+  def download(oid: ObjectId):GridFSDownloadObservable = {
+    gridfsBucket.downloadToObservable(oid)
+  }
+
+  def downloadToStream(oid: ObjectId, outputStream: OutputStream):Unit = {
+    val observable: GridFSDownloadObservable = gridfsBucket.downloadToObservable(oid)
+    observable.subscribe(GridFSStreamObserver(outputStream))
+  }
+
 
 }